chore: remove duplicate QuicChannelService and NgSctpService
This commit deduplicates the `QuicChannelService` and `NgSctpService` components which previously existed in both `libs/couch/src/commonMain/kotlin/borg/trikeshed/couch/ccek/` and `src/commonMain/kotlin/borg/trikeshed/ccek/`.
The implementation in `libs/couch` was identified as more complete (incorporating `QuicConnectionId`, `QuicShortFrame` for QUIC, and path failover logic for SCTP). Therefore, these comprehensive versions were moved to the canonical root location (`src/commonMain/kotlin/borg/trikeshed/ccek/`) and the redundant lesser implementations were removed. The package declaration was also correctly set to `borg.trikeshed.ccek.transport`.
